+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 1.0.1
 
-* TODO: Describe initial release.
+- Display region, province, city/municipality dropdown
+- Dependent display of list based on selection
+- Return string value of the selected item